requirements.txt
fastapi
uvicorn
openai
tiktoken
python-dotenv
requests
PyJWT
redis
SQLAlchemy
psycopg2-binary
pydantic
pydantic-settings
pytest
pika
pgvector
starlette
